Split post 1424 oversized slices
