+++ app/src/main/java/com/pocketshell/app/tmux/TmuxSessionReconnectTelemetry.kt
@@ -1,0 +1,31 @@
+package com.pocketshell.app.tmux
+
+import com.pocketshell.app.diagnostics.ReconnectCauseTrail
+import com.pocketshell.app.tmux.TmuxSessionViewModel.ConnectionStatus
+
+internal fun recordTmuxReconnectUiStateRendered(
+    status: ConnectionStatus,
+    hostId: Long,
+    canReconnect: Boolean,
+) {
+    when (status) {
+        is ConnectionStatus.Reconnecting -> ReconnectCauseTrail.record(
+            stage = "ui_reconnect_state",
+            outcome = "rendered",
+            cause = "connection_status_reconnecting",
+            "hostId" to hostId,
+            "attempt" to status.attempt,
+            "maxAttempts" to status.maxAttempts,
+            "retryDelayMs" to status.retryDelayMs,
+            "canReconnect" to canReconnect,
+        )
+        is ConnectionStatus.Failed -> ReconnectCauseTrail.record(
+            stage = "ui_reconnect_state",
+            outcome = "rendered",
+            cause = "connection_status_failed",
+            "hostId" to hostId,
+            "canReconnect" to canReconnect,
+        )
+        else -> Unit
+    }
+}
