+++ app/src/main/java/com/pocketshell/app/tmux/TmuxSessionViewModel.kt
@@ -17677,131 +17677,6 @@ public class TmuxSessionViewModel @Inject constructor(
     }
 
     /**
-     * EPIC #687 slice 1b: a VM-internal connection state whose vocabulary
-     * mirrors `:shared:core-connection`'s `ConnectionController.ConnectionState`
-     * (`Idle/Connecting/Attaching/Live/Backgrounded/Reattaching/Reconnecting/
-     * Unreachable/Gone`), so slice 1c can let `ConnectionController.state` drive
-     * [_connectionStatus] through the same [connectionStatusFor] mapper instead
-     * of being a body-swap rewrite.
-     *
-     * The view-facing [ConnectionStatus] becomes a pure PROJECTION of this
-     * state: every emission sets a [ConnectionState] and [connectionStatusFor]
-     * maps it to the exact [ConnectionStatus] the suite pins. To keep slice 1b a
-     * ZERO-behavior-change refactor (`ConnectionStatus` byte-identical to before),
-     * each state variant carries the full payload its projected [ConnectionStatus]
-     * needs. The unused-today variants ([Backgrounded], [Reattaching], [Gone]) are
-     * defined for 1c parity with the controller enum but are never emitted yet, so
-     * they change no current behavior.
-     *
-     * The §1 seam-table mapping (controller → status): `Connecting→Connecting`,
-     * `Attaching→Switching`, `Live→Connected`, `Reattaching/Reconnecting→
-     * Reconnecting`, `Unreachable→Failed`, `Gone→Failed`-equivalent, `Idle→Idle`.
-     * Deliberately, slice 1b does NOT simplify the `Failed`/"Tap Reconnect" band
-     * (that is the flagged slice-1c MIGRATE with maintainer sign-off); the mapper
-     * reproduces today's `Failed`/`Connected`/etc. payloads exactly.
-     */
-    private sealed interface ConnectionState {
-        data object Idle : ConnectionState
-
-        data class Connecting(val host: String, val port: Int, val user: String) :
-            ConnectionState
-
-        /** Warm same-host switch / warm open — projects to [ConnectionStatus.Switching]. */
-        data class Attaching(val host: String, val port: Int, val user: String) :
-            ConnectionState
-
-        /** Attached, input enabled — projects to [ConnectionStatus.Connected]. */
-        data class Live(val host: String, val port: Int, val user: String) :
-            ConnectionState
-
-        /**
-         * Controller-parity state for the backgrounded-but-warm window. Not emitted
-         * by the VM today (background detach keeps the prior status), so it is never
-         * mapped in practice in slice 1b; defined for 1c. Projects to the same
-         * [ConnectionStatus] as [Live] (the only sane projection if it ever flows).
-         */
-        data class Backgrounded(val host: String, val port: Int, val user: String) :
-            ConnectionState
-
-        /**
-         * Controller-parity silent in-grace reattach state. Not emitted by the VM
-         * today (the within-grace reattach path emits [Reconnecting]); defined for
-         * 1c. Projects to [ConnectionStatus.Reconnecting] like [Reconnecting].
-         */
-        data class Reattaching(
-            val host: String,
-            val port: Int,
-            val user: String,
-            val attempt: Int,
-            val maxAttempts: Int,
-            val retryDelayMs: Long,
-            val reason: String,
-        ) : ConnectionState
-
-        data class Reconnecting(
-            val host: String,
-            val port: Int,
-            val user: String,
-            val attempt: Int,
-            val maxAttempts: Int,
-            val retryDelayMs: Long,
-            val reason: String,
-        ) : ConnectionState
-
-        /**
-         * Controller-parity "target session deleted elsewhere" state (#666). Not
-         * emitted as a distinct status by the VM today; defined for 1c. Projects to
-         * [ConnectionStatus.Failed] (the `Gone→Failed`-equivalent §1 mapping).
-         */
-        data class Gone(val message: String) : ConnectionState
-
-        /** The honest error band — projects to [ConnectionStatus.Failed]. */
-        data class Unreachable(val message: String) : ConnectionState
-    }
-
-    /**
-     * EPIC #687 slice 1b: the PURE `ConnectionState → ConnectionStatus` projection
-     * from the §1 seam-reconciliation table. Reproduces every currently-pinned
-     * [ConnectionStatus] value BYTE-IDENTICALLY — the state variants carry the
-     * exact payloads, so this is a 1:1 facade map with no behavior change. Slice 1c
-     * will reuse this mapper to project `ConnectionController.state`.
-     */
-    private fun connectionStatusFor(state: ConnectionState): ConnectionStatus =
-        when (state) {
-            is ConnectionState.Idle -> ConnectionStatus.Idle
-            is ConnectionState.Connecting ->
-                ConnectionStatus.Connecting(state.host, state.port, state.user)
-            is ConnectionState.Attaching ->
-                ConnectionStatus.Switching(state.host, state.port, state.user)
-            is ConnectionState.Live ->
-                ConnectionStatus.Connected(state.host, state.port, state.user)
-            is ConnectionState.Backgrounded ->
-                ConnectionStatus.Connected(state.host, state.port, state.user)
-            is ConnectionState.Reattaching ->
-                ConnectionStatus.Reconnecting(
-                    host = state.host,
-                    port = state.port,
-                    user = state.user,
-                    attempt = state.attempt,
-                    maxAttempts = state.maxAttempts,
-                    retryDelayMs = state.retryDelayMs,
-                    reason = state.reason,
-                )
-            is ConnectionState.Reconnecting ->
-                ConnectionStatus.Reconnecting(
-                    host = state.host,
-                    port = state.port,
-                    user = state.user,
-                    attempt = state.attempt,
-                    maxAttempts = state.maxAttempts,
-                    retryDelayMs = state.retryDelayMs,
-                    reason = state.reason,
-                )
-            is ConnectionState.Gone -> ConnectionStatus.Failed(state.message)
-            is ConnectionState.Unreachable -> ConnectionStatus.Failed(state.message)
-        }
-
-    /**
      * EPIC #687 slice 1c-iv-a (THE STATUS FLIP): project the SHADOW
      * [ConnectionController]'s core [com.pocketshell.core.connection.ConnectionState]
      * onto the view-facing [ConnectionStatus] — the SINGLE source of the displayed
